refactoring specs; found conceptual problem with data model
I've determined that there's a conceptual problem in the code. I'm using the
term 'alarm' but it is not used anywhere in the specification. It is a remnant
of my initial thoughts that there should be a reusable, general purpose "alarm
scheduling subsystem" which would be used to schedule notifications. The idea is
not bad in itself but I've not managed to keep it encapsulated and independent
of the use case I'm solving and now I have this blurry concept of alarm all over
the code.

In the next couple of refactoring commits, I'll try to clean this up either by
abandoning the "alarm scheduling subsystem" idea or by implementing it properly.